Extract babel-loader rule and define tests for existing behavior
