+++ webpack.config.base.js
@@ -3,6 +3,7 @@ const webpack = require('webpack');
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const generateStripesAlias = require('./webpack/generate-stripes-alias');
+const babelLoaderRule = require('./webpack/babel-loader-rule');
 
 // React doesn't like being included multiple times as can happen when using
 // yarn link. Here we find a more specific path to it by first looking in
@@ -29,29 +30,7 @@ module.exports = {
   ],
   module: {
     rules: [
-      {
-        test(fn) {
-          // We want to transpile files inside node_modules/@folio or outside
-          // any node_modules directory. And definitely not files in
-          // node_modules outside the @folio namespace even if some parent
-          // directory happens to be in @folio.
-          //
-          // fn is the path after all symlinks are resolved so we need to be
-          // wary of all the edge cases yarn link will find for us.
-          const nmidx = fn.lastIndexOf('node_modules');
-          if (fn.endsWith('.js') && (nmidx === -1 || fn.lastIndexOf('@folio') > nmidx)) return true;
-          return false;
-        },
-        loader: 'babel-loader',
-        options: {
-          cacheDirectory: true,
-          presets: [
-            [require.resolve('babel-preset-env'), { modules: false }],
-            [require.resolve('babel-preset-stage-2')],
-            [require.resolve('babel-preset-react')],
-          ],
-        },
-      },
+      babelLoaderRule,
       {
         test: /\.json$/,
         loader: 'json-loader',
